Change quantity to integer

+++ database/migrations/2025_06_13_050734_create_expenses_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->decimal('amount', 10, 2);
             $table->date('date');
             $table->string('category')->nullable();
-            $table->decimal('quantity', 10, 2)->default(1);
+            $table->integer('quantity')->default(1);
             $table->string('unit')->nullable(); // e.g., 'pcs', 'kg', 'liters'
             $table->enum('payment_method', ['cash', 'e-wallet', 'check', 'bank_transfer', 'debit_card', 'credit_card'])->default('cash');
             $table->decimal('discount', 10, 2)->default(0.00); // Optional discount field
